coffee-packs/requirements.txt
fastapi==0.111.0
psycopg2-binary==2.9.9
py-consul==1.4.1
requests==2.31.0